+++ src/device/tsp.script/Script/ResourceManager.cs
@@ -1,4 +1,5 @@
 using System.Reflection;
+using cc.isr.Std.AssemblyExtensions;
 
 namespace cc.isr.VI.Tsp.Script;
 /// <summary>   Manager for resources. </summary>
@@ -6,6 +7,69 @@ namespace cc.isr.VI.Tsp.Script;
 [CLSCompliant( false )]
 public static class ResourceManager
 {
+
+    #region " create resources "
+
+    /// <summary>   Copies a file. </summary>
+    /// <remarks>
+    /// 2025-04-08. <para>
+    /// This is used for testing if a file can be copied while linked as a resource. </para>
+    /// </remarks>
+    /// <param name="fromPath"> Full pathname of from file. </param>
+    /// <param name="toPath">   Full pathname of to file. </param>
+    /// <param name="details">  [out] The details. </param>
+    /// <returns>   True if the test passes, false if the test fails. </returns>
+    public static bool CopyFile( string fromPath, string toPath, out string details )
+    {
+        if ( !System.IO.File.Exists( fromPath ) )
+        {
+            details = $"{fromPath} not found.";
+            return false;
+        }
+        else
+        {
+            System.IO.File.Copy( fromPath, toPath, true );
+            if ( !System.IO.File.Exists( toPath ) )
+            {
+                details = $"to file {toPath} not found; copy failed";
+                return false;
+            }
+            else
+            {
+                details = string.Empty;
+                return true;
+            }
+        }
+    }
+
+    /// <summary>   Writes a file. </summary>
+    /// <remarks>
+    /// 2025-04-08. <para>
+    /// This is used for testing if a file can be created while linked as a resource. </para>
+    /// </remarks>
+    /// <param name="toPath">   Full pathname of to file. </param>
+    /// <param name="contents"> The contents. </param>
+    /// <param name="details">  [out] The details. </param>
+    /// <returns>   True if the test passes, false if the test fails. </returns>
+    public static bool WriteFile( string toPath, string contents, out string details )
+    {
+        System.IO.File.WriteAllText( toPath, contents );
+        if ( !System.IO.File.Exists( toPath ) )
+        {
+            details = $"to file {toPath} not found; copy failed";
+            return false;
+        }
+        else
+        {
+            details = string.Empty;
+            return true;
+        }
+    }
+
+    #endregion
+
+    #region " binary script "
+
     /// <summary>   (Immutable) the binary script title. </summary>
     public const string BinaryScriptTitle = "isr_binaryScripts";
 
@@ -58,12 +122,9 @@ public static class ResourceManager
     /// <value> The script entity. </value>
     public static ScriptEntity? BinaryScriptEntity { get; set; } = null;
 
-    /// <summary>
-    /// Gets or sets the full pathname of the embedded resources files folder file.
-    /// </summary>
-    /// <value> The full pathname of the embedded resources files folder file. </value>
-    public static string EmbeddedResourcesFilesFolderPath { get; set; } = Path.Combine( typeof( ResourceManager ).Assembly.Location,
-        ResourceManager.EmbeddedResourceFolderName );
+    #endregion
+
+    #region " embedded resource assembly info "
 
     /// <summary>   (Immutable) pathname of the embedded resource folder. </summary>
     public const string EmbeddedResourceFolderName = "resources";
@@ -71,15 +132,30 @@ public static class ResourceManager
     /// <summary>   (Immutable) the assembly namespace. </summary>
     public const string AssemblyNamespace = "cc.isr.VI.Tsp";
 
+    #endregion
+
+    #region " read embedded resource "
+
+    /// <summary>   Builds full resource name. </summary>
+    /// <remarks>   2025-04-08. </remarks>
+    /// <param name="resourceFileName">     Filename of the resource file. </param>
+    /// <param name="resourceFolderName">   (Optional) ['resources'] name of the folder holding the resource. </param>
+    /// <returns>   A string. </returns>
+    public static string BuildFullResourceName( string resourceFileName, string resourceFolderName = "resources" )
+    {
+        return $"{ResourceManager.AssemblyNamespace}.{resourceFolderName}.{resourceFileName}";
+    }
+
     /// <summary>   Gets resource stream. </summary>
     /// <remarks>   2025-04-07. </remarks>
-    /// <param name="resourceName"> Name of the resource. </param>
+    /// <param name="resourceFileName">     Filename of the resource file. </param>
+    /// <param name="resourceFolderName">   (Optional) ['resources'] name of the folder holding the resource. </param>
     /// <returns>   The resource stream. </returns>
-    public static Stream? GetResourceStream( string resourceName )
+    public static Stream? GetResourceStream( string resourceFileName, string resourceFolderName = "resources" )
     {
-        // gets teh assembly that contains the code that is currently running.
+        // gets the assembly that contains the code that is currently running.
         Assembly? assembly = Assembly.GetExecutingAssembly();
-        return assembly.GetManifestResourceStream( resourceName );
+        return assembly.GetManifestResourceStream( ResourceManager.BuildFullResourceName( resourceFileName, resourceFolderName ) );
     }
 
     /// <summary>   Reads an embedded resource given the full resource path. </summary>
@@ -87,12 +163,12 @@ public static class ResourceManager
     /// <exception cref="FileNotFoundException">    Thrown when the requested file is not present. </exception>
     /// <param name="resourceName"> Name of the resource. </param>
     /// <returns>   The embedded resource content. </returns>
-    public static string ReadEmbeddedResource( string resourceName )
+    public static string ReadEmbeddedResourceFromFullName( string resourceName )
     {
         // gets teh assembly that contains the code that is currently running.
         Assembly? assembly = Assembly.GetExecutingAssembly();
         using Stream? stream = assembly.GetManifestResourceStream( resourceName )
-            ?? throw new FileNotFoundException( $"Embedded resource '{resourceName}' not found not found for the '{assembly.FullName}' assembly." );
+            ?? throw new FileNotFoundException( $"Embedded resource '{resourceName}' not found for the '{assembly.FullName}' assembly." );
         using StreamReader reader = new( stream );
         return reader.ReadToEnd();
     }
@@ -100,36 +176,50 @@ public static class ResourceManager
     /// <summary>   Reads embedded resource. </summary>
     /// <remarks>   2025-04-03. </remarks>
     /// <param name="resourceFileName">     Filename of the resource file. </param>
-    /// <param name="resourceFolderName">   (Optional) name of the folder holding the resource. </param>
+    /// <param name="resourceFolderName">   (Optional) ['resources'] name of the folder holding the resource. </param>
     /// <returns>   The embedded resource. </returns>
     public static string ReadEmbeddedResource( string resourceFileName, string resourceFolderName = "resources" )
     {
-        return ResourceManager.ReadEmbeddedResource( $"{ResourceManager.AssemblyNamespace}.{resourceFolderName}.{resourceFileName}" );
+        return ResourceManager.ReadEmbeddedResourceFromFullName( ResourceManager.BuildFullResourceName( resourceFileName, resourceFolderName ) );
     }
 
-    /// <summary>   Gets embedded resource file. </summary>
+    #endregion
+
+    #region " read resource file "
+
+    /// <summary>   Builds full resource file path. </summary>
+    /// <remarks>   2025-04-08. </remarks>
+    /// <param name="resourceFileName">     Filename of the resource file. </param>
+    /// <param name="resourceFolderName">   (Optional) ['resources'] name of the folder holding the resource. </param>
+    /// <returns>   A string. </returns>
+    public static string BuildResourceFilePath( string resourceFileName, string resourceFolderName = "resources" )
+    {
+        return Path.Combine( typeof( ResourceManager ).Assembly.DirectoryPath(), resourceFolderName, resourceFileName );
+    }
+
+    /// <summary>   Gets resource file info. </summary>
     /// <remarks>   2025-04-07. </remarks>
     /// <param name="resourceFileName">     Filename of the resource file. </param>
-    /// <param name="resourceFolderName">   (Optional) name of the folder holding the resource. </param>
-    /// <returns>   The embedded resource file. </returns>
-    public static FileInfo? GetEmbeddedResourceFile( string resourceFileName, string resourceFolderName = "resources" )
+    /// <param name="resourceFolderName">   (Optional) ['resources'] name of the folder holding the resource. </param>
+    /// <returns>   The resource file info. </returns>
+    public static FileInfo? GetResourceFileInfo( string resourceFileName, string resourceFolderName = "resources" )
     {
-        string filePath = Path.Combine( EmbeddedResourcesFilesFolderPath, resourceFolderName, resourceFileName );
+        string filePath = ResourceManager.BuildResourceFilePath( resourceFileName, resourceFolderName );
         if ( File.Exists( filePath ) )
             return new FileInfo( filePath );
         else
             return null;
     }
 
-    /// <summary>   Reads embedded resource file. </summary>
+    /// <summary>   Reads resource file. </summary>
     /// <remarks>   2025-04-07. </remarks>
     /// <exception cref="FileNotFoundException">    Thrown when the requested file is not present. </exception>
     /// <param name="resourceFileName">     Filename of the resource file. </param>
-    /// <param name="resourceFolderName">   (Optional) name of the folder holding the resource. </param>
+    /// <param name="resourceFolderName">   (Optional) ['resources'] name of the folder holding the resource. </param>
     /// <returns>   The embedded resource file. </returns>
-    public static string ReadEmbeddedResourceFile( string resourceFileName, string resourceFolderName = "resources" )
+    public static string ReadResourceFile( string resourceFileName, string resourceFolderName = "resources" )
     {
-        string filePath = Path.Combine( EmbeddedResourcesFilesFolderPath, resourceFolderName, resourceFileName );
+        string filePath = ResourceManager.BuildResourceFilePath( resourceFileName, resourceFolderName );
         if ( !File.Exists( filePath ) )
         {
             throw new FileNotFoundException( $"Embedded resource '{filePath}' not found." );
@@ -137,4 +227,5 @@ public static class ResourceManager
         return System.IO.File.ReadAllText( filePath );
     }
 
+    #endregion
 }