+++ changelog.md
@@ -2,6 +2,11 @@
 All notable changes to these libraries will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [8.1.9229] - 2025-04-08
+- Tsp.Script.ResourceManager
+  - Tighten code reading a resource file and an embedded resource file.
+  - Add methods to copy or create text file for testing the copy or creation of linked resources.
+
 ## [8.1.9228] - 2025-04-07
 - Add script file name builder.
 - Add resource manager class for loading the embedded resource.
@@ -252,6 +257,6 @@ which has issued a defect report.
 
 &copy; 2012 Integrated Scientific Resources, Inc. All rights reserved.
 
-[8.1.9228]: https://www.github.com/atecoder/dn.vi.ivi
+[8.1.9229]: https://www.github.com/atecoder/dn.vi.ivi
 [vs.VI]: https://www.github.com/atecoder/dn.vi.ivi
 [vs.Visa]: https://bitbucket.org/davidhary/vs.io.visa